+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- Polyline2D.tryFindSelfIntersection optimized
 
 ## [0.30.1] - 2026-03-24
 ### Fixed